api/requirements.txt
Flask
Flask-Bcrypt
Flask-Cors
Flask-JWT-Extended
Flask-Redis
Flask-SQLAlchemy
gunicorn
marshmallow
PyMySQL
pytest
python-dotenv